+++ metsim/cli/ms.py
@@ -26,6 +26,7 @@ import os
 import sys
 from collections import OrderedDict
 from configparser import ConfigParser
+from metsim import __name__, __version__, io
 
 
 def _is_valid_file(parser, arg):
@@ -38,7 +39,6 @@ def _is_valid_file(parser, arg):
 
 def parse(args):
     """Parse the command line arguments"""
-    from metsim import __name__, __version__
     parser = argparse.ArgumentParser()
     parser.add_argument('config', type=lambda x: _is_valid_file(parser, x),
                         help='Input configuration file')
@@ -54,61 +54,10 @@ def parse(args):
     return parser.parse_args(args)
 
 
-def init(opts):
-    """Initialize some information based on the options & config"""
-    config = ConfigParser()
-    config.optionxform = str
-    config.read(opts.config)
-    conf = OrderedDict(config['MetSim'])
-
-    def invert_dict(d):
-        return OrderedDict([reversed(item) for item in d.items()])
-
-    def to_list(s):
-        return json.loads(s.replace("'", '"').split('#')[0])
-
-    conf['forcing_vars'] = OrderedDict(config['forcing_vars'])
-    if conf['forcing_fmt'] != 'binary':
-        conf['forcing_vars'] = invert_dict(conf['forcing_vars'])
-    conf['domain_vars'] = invert_dict(OrderedDict(config['domain_vars']))
-    conf['state_vars'] = invert_dict(OrderedDict(config['state_vars']))
-    conf['chunks'] = OrderedDict(config['chunks'])
-    if 'constant_vars' in config:
-        conf['constant_vars'] = OrderedDict(config['constant_vars'])
-
-    # If the forcing variable is a directory, scan it for files
-    if os.path.isdir(conf['forcing']):
-        forcing_files = [os.path.join(conf['forcing'], fn) for fn in
-                         next(os.walk(conf['forcing']))[2]]
-    else:
-        forcing_files = conf['forcing']
-
-    # Ensure that parameters with boolean values are correctly recorded
-    for bool_param in ['utc_offset', 'period_ending']:
-        if (bool_param in conf.keys()
-            and conf[bool_param].strip().lower() == 'true'):
-            conf[bool_param] = True
-        else:
-            conf[bool_param] = False
-
-    # Update the full configuration
-    conf.update({"calendar": conf.get('calendar', 'standard'),
-                 "scheduler": opts.scheduler,
-                 "num_workers": opts.num_workers,
-                 "verbose": logging.DEBUG if opts.verbose else logging.INFO,
-                 "forcing": forcing_files,
-                 "out_dir": os.path.abspath(conf['out_dir']),
-                 "prec_type": conf.get('prec_type', 'uniform')})
-    conf['out_vars'] = to_list(conf.get('out_vars', '[]'))
-
-    conf = {k: v for k, v in conf.items() if v != []}
-    return conf
-
-
 def main():
     """Runs MetSim"""
     from metsim.metsim import MetSim
-    setup = init(parse(sys.argv[1:]))
+    setup = io.read_config(parse(sys.argv[1:]))
     ms = MetSim(setup)
     ms.run()